requirements.txt
Django==4.2.10
Pillow==10.1.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
python-decouple==3.8
whitenoise==6.6.0
django-ckeditor==6.7.0
django-taggit==5.0.1